+++ index.js
@@ -1,125 +1,107 @@
+// injector script based on https://github.com/joe27g/EnhancedDiscord/blob/beta/advanced_installation.md
+
 const path = require('path')
+const fs = require('fs-extra')
 const Module = require('module')
-const fs = require('fs')
 const electron = require('electron')
+const { BrowserWindow } = electron
 
 const conf = fs.existsSync(path.join(__dirname, 'config.json'))
   ? require(path.join(__dirname, 'config.json'))
   : {}
 const preloadPath = path.join(__dirname, 'engine')
 
-Object.assign(exports, {
-  inject (appPath) {
-    const basePath = path.join(appPath, '..', 'app.asar')
+process.env.DI_DEBUG_LOG = path.join(__dirname, 'debug.log')
+fs.removeSync(process.env.DI_DEBUG_LOG)
 
-    // fetch discord package.json
-    const pkg = require(path.join(basePath, 'package.json'))
+// patch browser window to use custom options
+class PatchedBrowserWindow extends BrowserWindow {
+  constructor (originalOptions) {
+    const options = Object.assign({}, originalOptions)
+    options.webPreferences = options.webPreferences || {}
 
-    // adjust electron root
-    electron.app.getAppPath = () => basePath
-
-    let splashPatched = false
-    let mainWindowPatched = false
-    let updaterPatched = false
-    let bootstrapPatched = false
-
-    // overwrite (and restore) the .js compiler
-    const oldLoader = Module._extensions['.js']
-    Module._extensions['.js'] = (mod, filename) => {
-      let content = fs.readFileSync(filename, 'utf8')
-      const fname = filename.toLowerCase()
-
-      // splash screen patches
-      if (fname.endsWith(`app_bootstrap${path.sep}splashscreen.js`)) {
-        splashPatched = true
-
-        content = content
-          // now add the real patch
-          .replace(
-            'new _electron.BrowserWindow(windowConfig);',
-            `new _electron.BrowserWindow(Object.assign(windowConfig, { webPreferences: { preload: "${path
-              .join(preloadPath, 'style.js')
-              .replace(/\\/g, '/')}" } }));
-
-              // disable csp
-              require("electron").session.defaultSession.webRequest.onHeadersReceived(function (details, done) {
-                const responseHeaders = {}
-                for (let k in details.responseHeaders) {
-                  if (!k.match(/^content-security/i)) {
-                    responseHeaders[k] = details.responseHeaders[k]
-                  }
-                }
-
-                done({ cancel: false, responseHeaders })
-              })
-              `
-          )
-
-        // main window patches
-      } else if (fname.endsWith(`mainscreen.js`)) {
-        mainWindowPatched = true
-
-        content = content
-          // transparency
-          .replace('transparent: false', `transparent: ${conf.transparent}`)
-        if (conf.transparent) {
-          content = content.replace(/backgroundColor: [^,]+,/, '')
-        }
-
-        // preload our script
-        if (content.match(/mainScreenPreload.js'\)/)) {
-          content = content.replace(
-            "mainScreenPreload.js')",
-            `mainScreenPreload.js'), preload: "${path
-              .join(preloadPath, 'index.js')
-              .replace(/\\/g, '/')}",`
-          )
-        } else {
-          content = content.replace(
-            'webPreferences: {',
-            `webPreferences: { preload: "${path
-              .join(preloadPath, 'index.js')
-              .replace(/\\/g, '/')}",`
-          )
-        }
-
-        if (!conf.transparent && typeof conf.frame === typeof true) {
-          // native frame
-          content = content
-            .replace('frame: false,', `frame: ${conf.frame},`)
-            .replace(
-              'mainWindowOptions.frame = true;',
-              `mainWindowOptions.frame = ${conf.frame};`
-            )
-        }
-      } else if (fname.endsWith(`app_bootstrap${path.sep}bootstrap.js`)) {
-        bootstrapPatched = true
-        if (conf.chromeFlags && Array.isArray(conf.chromeFlags)) {
-          const flags = JSON.stringify(
-            conf.chromeFlags.map(f => (Array.isArray(f) ? f : [f]))
-          )
-
-          content = content
-            // attach chrome flags
-            .replace(
-              'app.setVersion',
-              `${flags}.forEach(flag => app.commandLine.appendSwitch(flag[0], flag[1]));app.setVersion`
-            )
+    // options.center is only defined in the splash screen
+    if (!options.center) {
+      // add transparency support
+      if (conf.frame === true) {
+        options.frame = true
+      } else if (conf.frame === false) {
+        if (conf.transparency === true) {
+          options.transparent = true
+          delete options.backgroundColor
+        } else if (conf.transparency === false) {
+          options.transparent = false
         }
       }
-
-      if (
-        splashPatched &&
-        mainWindowPatched &&
-        updaterPatched &&
-        bootstrapPatched
-      ) {
-        Module._extensions['.js'] = oldLoader
-      }
-
-      return mod._compile(content, filename)
     }
 
-    Module._load(path.join(basePath, pkg.main), null, true)
+    process.env.DI_ORIG_PRELOAD = options.webPreferences.preload
+    options.webPreferences.preload = path.join(preloadPath, 'index.js')
+
+    return new BrowserWindow(options)
   }
+}
+
+const cacheEntry = Object.keys(require.cache)
+  .filter(k => k.match(/browser-window\.js$/i))
+  .pop()
+
+require.cache[cacheEntry].exports = PatchedBrowserWindow
+
+electron.app.on('ready', () => {
+  // patch webRequest session to ditch the CSP headers
+  electron.session.defaultSession.webRequest.onHeadersReceived(
+    (details, done) => {
+      // delete the content security response headers
+      Object.keys(details.responseHeaders)
+        .filter(k => k.toLowerCase() === 'content-security-policy')
+        .forEach(k => delete details.responseHeaders[k])
+
+      done(Object.assign(details, { cancel: false }))
+    }
+  )
+
+  // patch webRequest session to not ask for source map files
+  electron.session.defaultSession.webRequest.onBeforeRequest((details, done) =>
+    done({ cancel: details.url.endsWith('.js.map') })
+  )
 })
+
+// patch command line options
+if (conf.chromeFlags && Array.isArray(conf.chromeFlags)) {
+  conf.chromeFlags.forEach(flag => {
+    const key = Array.isArray(flag) ? flag[0] : flag
+    const value = Array.isArray(flag) ? flag[1] : null
+    electron.app.commandLine.appendSwitch(key, value)
+  })
+}
+
+exports.inject = function inject (appPath) {
+  const basePath = path.join(appPath, '..', 'app.asar')
+
+  // fetch discord package.json
+  const pkg = require(path.join(basePath, 'package.json'))
+
+  // adjust electron root
+  electron.app.getAppPath = () => basePath
+
+  /*
+  let updaterPatched = false
+
+  // overwrite (and restore) the .js compiler
+  const oldLoader = Module._extensions['.js']
+  Module._extensions['.js'] = (mod, filename) => {
+    let content = fs.readFileSync(filename, 'utf8')
+    // const fname = filename.toLowerCase()
+
+    if (updaterPatched) {
+      Module._extensions['.js'] = oldLoader
+    }
+
+    return mod._compile(content, filename)
+  }
+  */
+
+  // daisy chain into the original application file
+  Module._load(path.join(basePath, pkg.main), null, true)
+}